+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-## 1.10.2.dev (development stage/unreleased/unstable)
+## 1.10.3.dev (development stage/unreleased/unstable)
+
+## 1.10.3
 ### Changed
 - removed simplejson exception in restclient
 - set OSError from error to critical